# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PC Builder API",
    "contact": {
      "name": "Juozas Petryla"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/builds/{build_id}/components/{component_id}/offers": {
      "get": {
        "tags": [
          "Pardavėjų pasiūlymai"
        ],
        "summary": "Gauti komplekto komponento pasiūlymus",
        "description": "Trijų lygių hierarchija: komplektas → komponentas → pasiūlymai. Tikrinama priklausomybė.",
        "operationId": "listBuildComponentOffers",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          },
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Component Id"
            }
          },
          {
            "name": "in_stock",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "In Stock"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferRead"
                  },
                  "title": "Response Listbuildcomponentoffers"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Resursas nurodytoje hierarchijoje nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pardavėjų pasiūlymai"
        ],
        "summary": "Sukurti komponento pasiūlymą",
        "description": "Tik admin: prideda pasiūlymą matomo komplekto komponentui.",
        "operationId": "createBuildComponentOffer",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          },
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Component Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Resursas nurodytoje hierarchijoje nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Šio pardavėjo pasiūlymas jau yra.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/offers/{offer_id}": {
      "get": {
        "tags": [
          "Pardavėjų pasiūlymai"
        ],
        "summary": "Gauti komponento pasiūlymą",
        "description": "Grąžina pasiūlymą pagal unikalų ID. Komponentas ir komplektas pasiekiami per links.",
        "operationId": "getOffer",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Resursas nurodytoje hierarchijoje nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pardavėjų pasiūlymai"
        ],
        "summary": "Atnaujinti komponento pasiūlymą",
        "description": "Tik admin: pakeičia pasiūlymą matomame komplekte, nekeisdamas jo komponento.",
        "operationId": "replaceOffer",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Resursas nurodytoje hierarchijoje nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Šio pardavėjo pasiūlymas jau yra.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pardavėjų pasiūlymai"
        ],
        "summary": "Pašalinti komponento pasiūlymą",
        "description": "Tik admin: pašalina pasiūlymą matomame komplekte; komponentas ir komplektas nešalinami.",
        "operationId": "deleteOffer",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Resursas nurodytoje hierarchijoje nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OfferCreate": {
        "properties": {
          "retailer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Retailer",
            "examples": [
              "Kilobaitas"
            ]
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Price",
            "description": "Teigiama kaina, iki 2 skaitmenų po kablelio. Atsakyme — JSON eilutė.",
            "examples": [
              "299.99"
            ]
          },
          "product_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "format": "uri",
            "title": "Product Url",
            "examples": [
              "https://example.com/products/ryzen-7-7800x3d"
            ]
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "retailer",
          "price",
          "product_url"
        ],
        "title": "OfferCreate"
      },
      "OfferLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "component": {
            "type": "string",
            "title": "Component"
          },
          "build": {
            "type": "string",
            "title": "Build"
          }
        },
        "type": "object",
        "required": [
          "self",
          "component",
          "build"
        ],
        "title": "OfferLinks"
      },
      "OfferRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "component_id": {
            "type": "integer",
            "title": "Component Id"
          },
          "retailer": {
            "type": "string",
            "title": "Retailer"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "product_url": {
            "type": "string",
            "title": "Product Url"
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock"
          },
          "links": {
            "$ref": "#/components/schemas/OfferLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "component_id",
          "retailer",
          "price",
          "product_url",
          "in_stock",
          "links"
        ],
        "title": "OfferRead"
      },
      "OfferReplace": {
        "properties": {
          "retailer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Retailer",
            "examples": [
              "Kilobaitas"
            ]
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Price",
            "description": "Teigiama kaina, iki 2 skaitmenų po kablelio. Atsakyme — JSON eilutė.",
            "examples": [
              "299.99"
            ]
          },
          "product_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "format": "uri",
            "title": "Product Url",
            "examples": [
              "https://example.com/products/ryzen-7-7800x3d"
            ]
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "retailer",
          "price",
          "product_url"
        ],
        "title": "OfferReplace"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}